Move addExports to its own file
